css-fix i chat och för knappar, samt lite utbrytande av projekt. Skapade komponent Joke
