+++ src/pages/Project/Chat/Chat.module.css
@@ -1,7 +1,7 @@
 .container{
   width: 100%;
   height: 25vh;
-  padding-top: 10px;
+  padding: 20px;
 }
 .nameRight{
   margin-right: 17px;
@@ -13,7 +13,6 @@
   font-size: 12px;
   color: var(--color-grey-light);
 }
-
 .messageRight {
   text-align:right;
   width:100%;
@@ -26,18 +25,71 @@
 .bubbleRight{
   background:var(--color-primary-dark);
   width:60%;
-  border-radius: 10px;
-  margin-left: 37.5%;
+  border-radius: 12px;
+  margin-left: 38.2%;
 }
 .bubbleLeft{
-  background:var(--color-primary-dark);
+  background:var(--color-primary-ultra-dark);
   width:60%;
-  border-radius: 10px;
+  border-radius: 12px;
   margin-left: 10px;
 }
-.chattText{
+.chatText{
   margin: 0 5px 0 5px;
   padding:5px;
 }
+.chatSquare{
+  width:75%;
+  margin-left: 12%;
+}
+.chatBackground{
+  width:100%;
+  height: 70vh;
+  overflow: auto;
+  background:var(--color-grey-dark);
+  border-radius: 20px;
+  text-align: right;
+}
+.clearButton{
+  color:var(--color-grey-light);
+  margin: 10px 10px 5px 10px;
+  padding:5px 10px 5px 10px;
+  background: var(--color-primary-ultra-dark);
+  border: none;
+  border-radius: 12px;
+  box-shadow: 2px 2px 12px -4px rgba(35,35,35,0.2);
+}
+.clearButton:hover{
+  cursor: pointer;
+  opacity: 0.8;
+}
+.textInput{
+  width:81%;
+  height: 2rem;
+  padding:10px;
+  border-radius: 12px;
+  border:none;
+  margin: 5px;
+}
+.textInput:focus{
+  outline:none;
+}
+.chatButton{
+  background: var(--color-primary-medium);
+  border-radius: 16px;
+  width:fit-content;
+  border:none;
+  height: 2rem;
+  box-shadow: 5px 5px 12px -4px rgba(35,35,35,0.4);
+  color:var(--color-almost-black);
+  font-size: 16px;
+  margin: 3px;
+  padding:5px 20px 5px 20px;
+}
+.chatButton:hover{
+  cursor: pointer;
+  opacity: 0.8;
+}
+
