+++ src/ui-lib/Button/Button.module.css
@@ -7,6 +7,7 @@
     padding: 0 32px;
     border-radius: 30px;
     background: var(--color-primary-medium);
+    box-shadow: 5px 5px 12px -4px rgba(35,35,35,0.4);
   }
   
   .container:hover {
@@ -15,7 +16,7 @@
   }
   
   .text {
-    color: var(--color-grey-ultra-dark);
+    color: var(--color-almost-black);
     font-size: 16px;
     line-height: 150%;
     font-weight: 500;
