+++ src/ui-lib/Divider/SpacerDivider.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Spacer } from "../Spacer/Spacer";
+import { Divider } from "./Divider";
+
+export const SpacerDivider =()=>{
+    return(<>
+    <Spacer size={16}/>
+    <Divider/>
+    <Spacer size={16}/>
+    </>)
+}
